:recycle: Restructure motion imports
